+++ app/src/main/java/com/github/dfa/diaspora_android/util/DiasporaUrlHelper.java
@@ -18,7 +18,10 @@
  */
 package com.github.dfa.diaspora_android.util;
 
+import com.github.dfa.diaspora_android.App;
+import com.github.dfa.diaspora_android.R;
 import com.github.dfa.diaspora_android.data.AppSettings;
+import com.github.dfa.diaspora_android.data.PodAspect;
 
 /**
  * Helper class that provides easy access to specific urls related to diaspora
@@ -54,140 +57,175 @@ public class DiasporaUrlHelper {
     /**
      * Return a https url of the pod set in AppSettings.
      * Eg. https://pod.geraspora.de
+     *
      * @return https://(pod-domain.tld)
      */
     public String getPodUrl() {
-        return HTTPS+settings.getPodDomain();
+        return HTTPS + settings.getPodDomain();
     }
 
     /**
      * Return a https url that points to the stream of the configured diaspora account
+     *
      * @return https://(pod-domain.tld)/stream
      */
     public String getStreamUrl() {
-        return getPodUrl()+SUBURL_STREAM;
+        return getPodUrl() + SUBURL_STREAM;
     }
 
     /**
      * Return a https url that points to the notifications feed of the configured diaspora account
+     *
      * @return https://(pod-domain.tld)/notifications
      */
     public String getNotificationsUrl() {
-        return getPodUrl()+SUBURL_NOTIFICATIONS;
+        return getPodUrl() + SUBURL_NOTIFICATIONS;
     }
 
     /**
      * Returns a https url that points to the post with the id postId
+     *
      * @return https://(pod-domain.tld)/posts/(postId)
      */
     public String getPostsUrl(long postId) {
-        return getPodUrl()+SUBURL_POSTS+postId;
+        return getPodUrl() + SUBURL_POSTS + postId;
     }
 
     /**
      * Return a https url that points to the conversations overview of the registered diaspora account
+     *
      * @return https://(pod-domain.tld)/conversations
      */
     public String getConversationsUrl() {
-        return getPodUrl()+SUBURL_CONVERSATIONS;
+        return getPodUrl() + SUBURL_CONVERSATIONS;
     }
 
     /**
      * Return a https url that points to the new-post form that lets the user create a new post
+     *
      * @return https://(pod-domain.tld)/status_messages/new
      */
     public String getNewPostUrl() {
-        return getPodUrl()+SUBURL_NEW_POST;
+        return getPodUrl() + SUBURL_NEW_POST;
     }
 
     /**
      * Return a https url that shows the profile of the currently registered diaspora account
+     *
      * @return https://(pod-domain.tld)/people/(profileId)
      */
     public String getProfileUrl() {
-        return getPodUrl()+SUBURL_PEOPLE+settings.getProfileId();
+        return getPodUrl() + SUBURL_PEOPLE + settings.getProfileId();
     }
 
     /**
      * Return a https url that shows the profile of the user with user id profileId
+     *
      * @param profileId Id of the profile to be shown
      * @return https://(pod-domain.tld)/people/(profileId)
      */
     public String getProfileUrl(long profileId) {
-        return getPodUrl()+SUBURL_PEOPLE+profileId;
+        return getPodUrl() + SUBURL_PEOPLE + profileId;
     }
 
     /**
      * Return a https url that points to the activities feed of the currently registered diaspora account
+     *
      * @return https://(pod-domain.tld)/activity
      */
     public String getActivityUrl() {
-        return getPodUrl()+SUBURL_ACTIVITY;
+        return getPodUrl() + SUBURL_ACTIVITY;
     }
 
     /**
      * Return a https url that points to the feed of posts that were liked by the currently registered diaspora account
+     *
      * @return https://(pod-domain.tld)/liked
      */
     public String getLikedPostsUrl() {
-        return getPodUrl()+SUBURL_LIKED;
+        return getPodUrl() + SUBURL_LIKED;
     }
 
     /**
      * Return a https url that points to the stream of posts that were commented by the currently registered diaspora account
+     *
      * @return https://(pod-domain.tld)/commented
      */
     public String getCommentedUrl() {
-        return getPodUrl()+SUBURL_COMMENTED;
+        return getPodUrl() + SUBURL_COMMENTED;
     }
 
     /**
      * Return a https url that points to the stream of posts in which the currently registered diaspora account has been mentioned in
+     *
      * @return https://(pod-domain.tld)/mentions
      */
     public String getMentionsUrl() {
-        return getPodUrl()+SUBURL_MENTIONS;
+        return getPodUrl() + SUBURL_MENTIONS;
     }
 
     /**
      * Return a https url that points to the stream of public posts
+     *
      * @return https://(pod-domain.tld)/public
      */
     public String getPublicUrl() {
-        return getPodUrl()+SUBURL_PUBLIC;
+        return getPodUrl() + SUBURL_PUBLIC;
     }
 
     /**
      * Return a https url that toggles between mobile and desktop view when opened
+     *
      * @return https://(pod-domain.tld)/mobile/toggle
      */
     public String getToggleMobileUrl() {
-        return getPodUrl()+SUBURL_TOGGLE_MOBILE;
+        return getPodUrl() + SUBURL_TOGGLE_MOBILE;
     }
 
     /**
      * Return a https url that queries posts for the given hashtag query
+     *
      * @param query hashtag to be searched
      * @return https://(pod-domain.tld)/tags/query
      */
     public String getSearchTagsUrl(String query) {
-        return getPodUrl()+SUBURL_SEARCH_TAGS+query;
+        return getPodUrl() + SUBURL_SEARCH_TAGS + query;
     }
 
     /**
      * Return a https url that queries user accounts for query
+     *
      * @param query search term
      * @return https://(pod-domain.tld)/people.mobile?q=(query)
      */
     public String getSearchPeopleUrl(String query) {
-        return getPodUrl()+SUBURL_SEARCH_PEOPLE+query;
+        return getPodUrl() + SUBURL_SEARCH_PEOPLE + query;
     }
 
     /**
      * Returns the url of the blank WebView
+     *
      * @return about:blank
      */
     public String getBlankUrl() {
         return URL_BLANK;
+    }
+
+    public boolean isAspectUrl(String url) {
+        return url.startsWith(getPodUrl() + "/aspects?a_ids[]=");
+    }
+
+    public String getAspectNameFromUrl(String url, App app) {
+        url = url.replace(getPodUrl() + "/aspects?a_ids[]=", "").split(",")[0];
+        try {
+            int id = Integer.parseInt(url);
+            for (PodAspect aspect : app.getPodUserProfile().getAspects()) {
+                if (aspect.id == id) {
+                    return aspect.name;
+                }
+            }
+        } catch (Exception ignored) {
+        }
+        return app.getString(R.string.aspects);
     }
 }